refactor: rename MovieDetailsDescription to MovieDescription
