+++ src/components/MovieDescription/MovieDescription.scss
