+++ src/components/MovieDescription/index.js
@@ -1,0 +1,3 @@
+import MovieDescription from './MovieDescription';
+
+export default MovieDescription;
