--- src/components/MovieDetailsDescription/index.js
@@ -1,3 +1,0 @@
-import MovieDetailsDescription from './MovieDetailsDescription';
-
-export default MovieDetailsDescription;
